SD_Game/requirements.txt
PyQt6>=6.6.0
pandas>=2.2.0
numpy>=1.26.0
PyYAML>=6.0.1
pyqtgraph>=0.13.3